response: extract usage reports and related types to their own module
While at it also add a couple of unit tests.
