+++ src/response.rs
@@ -1,12 +1,8 @@
 use std::prelude::v1::*;
 
-use std::{fmt, str::FromStr};
+use std::str::FromStr;
 
-use chrono::prelude::*;
-use serde::{
-    de::{self, Deserializer, MapAccess, Visitor},
-    Deserialize,
-};
+use serde::Deserialize;
 
 mod app_keys_list;
 pub use app_keys_list::ListAppKeys;
@@ -14,123 +10,8 @@ pub use app_keys_list::ListAppKeys;
 mod metrics_hierarchy;
 pub use metrics_hierarchy::MetricsHierarchy;
 
-mod systemtime {
-    use chrono::DateTime;
-
-    #[repr(transparent)]
-    #[derive(Debug, Clone, PartialEq, Eq)]
-    pub struct PeriodTime(pub i64);
-
-    impl<Tz: chrono::TimeZone> From<DateTime<Tz>> for PeriodTime {
-        fn from(dt: DateTime<Tz>) -> PeriodTime {
-            PeriodTime(dt.timestamp())
-        }
-    }
-}
-
-pub use systemtime::PeriodTime;
-
-#[non_exhaustive]
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub enum UsageReportError {
-    LimitsExceeded(u64),
-    Overflow,
-}
-
-#[derive(Debug, Clone, Deserialize, PartialEq, Eq)]
-#[serde(rename = "usage_report")]
-pub struct UsageReport {
-    pub metric: String,
-    pub period: Period,
-    pub period_start: PeriodTime,
-    pub period_end: PeriodTime,
-    pub max_value: u64,
-    pub current_value: u64,
-}
-
-impl UsageReport {
-    pub fn metric(&self) -> &str {
-        self.metric.as_ref()
-    }
-
-    pub fn period(&self) -> &Period {
-        &self.period
-    }
-
-    pub fn period_times(&self) -> (&PeriodTime, &PeriodTime) {
-        (&self.period_start, &self.period_end)
-    }
-
-    pub fn max_value(&self) -> u64 {
-        self.max_value
-    }
-
-    pub fn current_value(&self) -> u64 {
-        self.max_value
-    }
-
-    pub fn remaining(&self) -> u64 {
-        self.max_value.checked_sub(self.current_value).unwrap_or(0)
-    }
-
-    pub fn is_limited(&self) -> bool {
-        self.current_value >= self.max_value
-    }
-
-    pub fn authorize(&self, hits: u64) -> Result<u64, UsageReportError> {
-        let new_hits = self
-            .current_value
-            .checked_add(hits)
-            .ok_or(UsageReportError::Overflow)?;
-        if new_hits > self.max_value {
-            Err(UsageReportError::LimitsExceeded(new_hits - self.max_value))
-        } else {
-            Ok(new_hits)
-        }
-    }
-
-    pub fn report(&mut self, hits: u64) -> Result<u64, UsageReportError> {
-        self.current_value = self
-            .current_value
-            .checked_add(hits)
-            .ok_or(UsageReportError::Overflow)?;
-
-        Ok(self.current_value)
-    }
-
-    pub fn reset<V: Into<Option<u64>>>(&mut self, val: V) {
-        self.current_value = val.into().unwrap_or(0);
-    }
-}
-
-// Unfortunately the XML output from Apisonator includes a rather useless "usage_reports" tag that
-// is then followed by a "usage_report" tag in each UsageReport, so we need to wrap that up.
-#[cfg_attr(supports_transparent_enums, repr(transparent))]
-#[derive(Debug, Clone, Deserialize, PartialEq, Eq)]
-pub enum UsageReports {
-    #[serde(rename = "usage_report")]
-    UsageReports(Vec<UsageReport>),
-}
-
-impl UsageReports {
-    pub fn as_vec(&self) -> &Vec<UsageReport> {
-        match self {
-            Self::UsageReports(v) => v,
-        }
-    }
-
-    pub fn as_vec_mut(&mut self) -> &mut Vec<UsageReport> {
-        match self {
-            Self::UsageReports(v) => v,
-        }
-    }
-
-    pub fn into_inner(self) -> Vec<UsageReport> {
-        match self {
-            Self::UsageReports(v) => v,
-        }
-    }
-}
+mod usage_report;
+pub use usage_report::{Period, PeriodTime, UsageReport, UsageReportError, UsageReports};
 
 #[derive(Debug, Clone, Deserialize, PartialEq, Eq)]
 #[serde(rename_all = "snake_case")]
@@ -243,94 +124,6 @@ impl AuthorizationError {
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum Period {
-    Minute,
-    Hour,
-    Day,
-    Week,
-    Month,
-    Year,
-    Eternity,
-    Other(String),
-}
-
-struct PeriodStringVisitor;
-
-impl<'de> Visitor<'de> for PeriodStringVisitor {
-    type Value = Period;
-
-    fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-        formatter.write_str("a string that represents a period")
-    }
-
-    fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
-    where
-        E: de::Error,
-    {
-        match v {
-            "minute" => Ok(Period::Minute),
-            "hour" => Ok(Period::Hour),
-            "day" => Ok(Period::Day),
-            "week" => Ok(Period::Week),
-            "month" => Ok(Period::Month),
-            "year" => Ok(Period::Year),
-            "eternity" => Ok(Period::Eternity),
-            period_name => Ok(Period::Other(period_name.into())),
-        }
-    }
-}
-
-impl<'de> Deserialize<'de> for Period {
-    fn deserialize<D>(deserializer: D) -> Result<Period, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        deserializer.deserialize_any(PeriodStringVisitor)
-    }
-}
-
-struct TimestampVisitor;
-
-impl<'de> Visitor<'de> for TimestampVisitor {
-    type Value = PeriodTime;
-
-    fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-        formatter.write_str("a string that represents a timestamp")
-    }
-
-    fn visit_map<V>(self, mut map: V) -> Result<PeriodTime, V::Error>
-    where
-        V: MapAccess<'de>,
-    {
-        // We know there's only one key with one value.
-        // The key is not used, but we need to call "next_key()". From the
-        // docs: "Calling `next_value` before `next_key` is incorrect and is
-        // allowed to panic or return bogus results.".
-        let _key: Option<String> = map.next_key()?;
-        let timestamp: String = map.next_value()?;
-
-        let ts_str = timestamp.as_str();
-        let dt = DateTime::parse_from_str(ts_str, "%Y-%m-%d %H:%M:%S %z").map_err(|e| {
-            de::Error::custom(format_args!(
-                "invalid timestamp {}, expected %Y-%m-%d %H:%M:%S %z: {:?}",
-                ts_str, e
-            ))
-        })?;
-
-        Ok(PeriodTime::from(dt))
-    }
-}
-
-impl<'de> Deserialize<'de> for PeriodTime {
-    fn deserialize<D>(deserializer: D) -> Result<PeriodTime, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        deserializer.deserialize_any(TimestampVisitor)
-    }
-}
-
 #[repr(transparent)]
 #[derive(Debug, Deserialize, PartialEq)]
 pub struct Metric(pub String);
@@ -345,7 +138,8 @@ impl FromStr for Authorization {
 
 #[cfg(test)]
 mod tests {
-    use super::{UsageReports::*, *};
+    use super::{UsageReports::UsageReports, *};
+    use chrono::prelude::*;
 
     #[test]
     fn parse() {
